backend/requirements.txt
fastapi>=0.115,<0.116
uvicorn[standard]>=0.34,<0.35
pydantic-settings>=2.8,<2.9
python-multipart>=0.0.20,<0.0.21
httpx>=0.28,<0.29
Pillow>=10.1,<13
numpy>=1.26,<2
torch>=2.0,<2.2
torchvision>=0.15,<0.17
fvcore>=0.1.5.post20221221,<0.2
einops>=0.8,<0.9
prettytable>=3.17,<4
scikit-image>=0.25,<0.26
opencv-python>=4.8,<5
